refactor: Extract config adjusters
